Remove unneeded tower service changes
