+++ crates/rover-client/src/operations/graph/introspect/mod.rs
@@ -1,9 +1,7 @@
 mod runner;
 mod schema;
-mod service;
 mod types;
 
 pub use runner::run;
 pub use schema::Schema;
-pub use service::{GraphIntrospect, GraphIntrospectError};
 pub use types::{GraphIntrospectInput, GraphIntrospectResponse};
