+++ crates/rover-client/src/operations/graph/introspect/types.rs
@@ -1,8 +1,7 @@
 use std::collections::HashMap;
 
-use crate::operations::graph::introspect::service::graph_introspect_query;
+use crate::operations::graph::introspect::runner::graph_introspect_query;
 
-#[cfg(test)]
 pub(crate) type QueryResponseData = graph_introspect_query::ResponseData;
 pub(crate) type QueryVariables = graph_introspect_query::Variables;
