+++ crates/rover-client/src/operations/supergraph/fetch/mod.rs
@@ -1,7 +1,5 @@
 mod runner;
-mod service;
 mod types;
 
 pub use runner::run;
-pub use service::{SupergraphFetch, SupergraphFetchRequest};
 pub use types::SupergraphFetchInput;
